requirements.txt
streamlit>=1.30.0
textstat>=0.7.0
fpdf2>=2.7.0
pandas>=2.0.0
